Undo see_ge changes
The subexpression:

gating_type(m) == HAWK ? attacks_bb<HAWK>(from, occupied)
                       : attacks_bb<ELEPHANT>(from, occupied);

always evaluates to true. I suppose that an "& to" was intended to appear
but it doesn't.

Effectively this means that when we consider SEE sequences, the opponent
king is never allowed to make a capture if the first move was a gating
move and when we make a gating move ourselves, should our king ever get
involved in the sequence then we can substitue our gated piece even if
it doesn't attack the destination square.

One alternative is to add in the "& to" that was missing. Instead I am
proposing to get rid of the whole thing and run a SPRT(-10, 0). We can
start from a blank slate if this works.